--- test/matplotex/line_plot_test.exs
@@ -1,3 +1,0 @@
-defmodule Matplotex.LinePlotTest do
-  use Matplotex.PlotCase
-end
